refactor: cut out filter/sort UI into a component
